Refactor user result and people result
